Revert " Handle case when story is moved projects"
