+++ server/server/controllers/webhook/cases/index.js
@@ -1,18 +1,16 @@
-const storyCreated = require('./story-created');
-const storyRenamed = require('./story-renamed');
-const storyArchived = require('./story-Archived');
-const storyUnarchived = require('./story-unarchived');
-const storyChangedProject = require('./story-changed-project');
-const storyMovedToVisibleColumn = require('./story-moved-to-visible-column');
-const storyMovedFromVisibleColumn = require('./story-moved-from-visible-column');
-const storyMovedBetweenVisibleColumns = require('./story-moved-between-visible-columns');
+const storyCreated = require('./storyCreated');
+const storyRenamed = require('./storyRenamed');
+const storyArchived = require('./storyArchived');
+const storyUnarchived = require('./storyUnarchived');
+const storyMovedToVisibleColumn = require('./storyMovedToVisibleColumn');
+const storyMovedFromVisibleColumn = require('./storyMovedFromVisibleColumn');
+const storyMovedBetweenVisibleColumns = require('./storyMovedBetweenVisibleColumns');
 
 module.exports = {
   storyCreated,
   storyRenamed,
   storyArchived,
   storyUnarchived,
-  storyChangedProject,
   storyMovedToVisibleColumn,
   storyMovedFromVisibleColumn,
   storyMovedBetweenVisibleColumns,
