+++ server/server/controllers/webhook/cases/storyMovedBetweenVisibleColumns.js
