+++ server/server/controllers/webhook/cases/storyMovedFromVisibleColumn.js
