+++ server/server/controllers/webhook/cases/storyMovedToVisibleColumn.js
